+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -4,6 +4,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.UserStorage;
 
@@ -22,17 +23,17 @@ public class UserService {
     }
 
     public List<Optional<User>> getUsers() {
+        if (userStorage.getUsers().isEmpty()) {
+            return null;
+        }
         return userStorage.getUsers();
     }
 
     public User create(User user) {
-
         if (user.getName().equals("")) {
             user.setName(user.getLogin());
             log.info("Пустое имя пользователя заменено на Login.");
         }
-        //userDao.create(user);
-        //getStorage().put(user.getId(), user);
         log.info("Создан пользователь.");
         return userStorage.create(user);
     }
@@ -42,35 +43,55 @@ public class UserService {
             user.setName(user.getLogin());
             log.info("Пустое имя пользователя заменено на Login.");
         }
-        //userDao.create(user);
-        //getStorage().put(user.getId(), user);
-        log.info("Создан пользователь.");
-        //return userStorage.create(user);
-
-        return userStorage.update(user);
+        log.info("Пользователь обновлен.");
+        if (userStorage.findUserByStorage(user.getId())) {
+            log.info("User id - {}", user.getId());
+            log.info("User id in base - {}", getUser(user.getId()).get().getId());
+            return userStorage.update(user);
+        } else {
+            log.info("Пользователь с идентификатором {} не найден.", user.getId());
+            throw new NotFoundException("Такого пользователя нет.");
+        }
     }
 
     public Optional<User> getUser(Long id) {
-        return userStorage.getUser(id);
+        if (userStorage.findUserByStorage(id)) {
+            return userStorage.getUser(id);
+        } else {
+            log.info("Пользователь с идентификатором {} не найден.", id);
+            throw new NotFoundException("Такого пользователя нет.");
+        }
     }
 
     //Добавление в друзья
     public void addFriends(Long id, Long friendId) {
-userStorage.addFriends(id, friendId);
+        if (userStorage.findUserByStorage(id) && userStorage.findUserByStorage(friendId)) {
+            userStorage.addFriends(id, friendId);
+        } else {
+            log.info("Пользователь не найден.");
+            throw new NotFoundException("Пользователь не найден.");
+        }
     }
 
-    public List<Optional<User>> findAllUserFriends(Long userId){
-
+    public List<Optional<User>> findAllUserFriends(Long userId) {
         return userStorage.findAllUserFriends(userId);
     }
 
     //Удалить пользователя из друзей
-    public void deleteFriend(Long id, Long fid){
-        userStorage.deleteFriend(id, fid);
+    public void deleteFriend(Long id, Long fid) {
+        if (userStorage.findUserByStorage(id) && userStorage.findUserByStorage(fid)) {
+            userStorage.deleteFriend(id, fid);
+        } else {
+            log.info("Друг - {} - не найден.", fid);
+            throw new NotFoundException("Пользователь не найден.");
+        }
     }
 
     public List<Optional<User>> commonFriends(Long id, Long otherId) {
-        return userStorage.commonFriends(id, otherId);
+        List<Optional<User>> userId = new ArrayList<>(findAllUserFriends(id));
+        List<Optional<User>> userOtherId = new ArrayList<>(findAllUserFriends(otherId));
+        userId.retainAll(userOtherId);
+        return userId;
     }
 
 }
